Added tests to commit. 64 bit simx still not working

+++ sim/simX/types.h
@@ -7,6 +7,7 @@
 namespace vortex {
 
 typedef uint8_t  Byte;
+// simx64
 typedef uint64_t Word;
 typedef int64_t  WordI;
 
@@ -15,8 +16,8 @@ typedef uint32_t HalfWord;
 typedef int32_t HalfWordI;
 
 // simx64
-typedef uint32_t Addr;
-typedef uint32_t Size;
+typedef uint64_t Addr;
+typedef uint64_t Size;
 
 typedef std::bitset<32> RegMask;
